DIRECTORY-193 : Correction of sonar issues.
git-svn-id: http://dev.lutece.paris.fr/svn/lutece/portal/trunk/plugins/directory-form-workflow/plugin-directory@49451 bab10101-e421-0410-a517-8ce0973de3ef

+++ src/java/fr/paris/lutece/plugins/directory/web/DoDownloadFile.java
@@ -139,7 +139,7 @@ public class DoDownloadFile
 					strMimeType = FileSystemUtil.getMIMEType( file.getTitle( ) );
 				}
 				response.setContentType( strMimeType );
-                response.setContentLength( (int) byteFileOutPut.length );
+                response.setContentLength( byteFileOutPut.length );
 
                 OutputStream os = response.getOutputStream(  );
                 os.write( byteFileOutPut );
